Quick fix, cleanup
