+++ .idea/sunback_project.iml
@@ -10,7 +10,7 @@
       <excludeFolder url="file://$MODULE_DIR$/sunback/sunback_images" />
       <excludeFolder url="file://$MODULE_DIR$/venv2" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.8 (newtest)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.8 (sunback_env)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
   <component name="PackageRequirementsSettings">
